requirements.txt
# Для работы с MySQL
pymysql           # Драйвер для подключения к MySQL

# Для работы с MongoDB
pymongo           # Драйвер для подключения к MongoDB

# Для загрузки переменных окружения из .env
python-dotenv     # Позволяет использовать os.getenv с .env-файлом

# Для красивого вывода таблиц в консоли
tabulate          # Форматирование таблиц для CLI

flake8            # Линтер для проверки кода на соответствие PEP 8
black             # Автоматическое форматирование кода по PEP 8